requirements.txt
discord.py
pynacl
opuslib